Added tag converter class
