--- cs/Markdown/Html/BoldTag.cs
@@ -1,11 +1,0 @@
-﻿using Markdown.Tokens;
-
-namespace Markdown.Html;
-
-public class BoldTag : IHtmlTag
-{
-    public Token ToHtml(string value)
-    {
-        throw new NotImplementedException();
-    }
-}
