--- cs/Markdown/Html/IHtmlTag.cs
@@ -1,8 +1,0 @@
-﻿using Markdown.Tokens;
-
-namespace Markdown.Html;
-
-public interface IHtmlTag
-{
-    Token ToHtml(string value);
-}
