--- cs/Markdown/Html/ItalicTag.cs
@@ -1,11 +1,0 @@
-﻿using Markdown.Tokens;
-
-namespace Markdown.Html;
-
-public class ItalicTag : IHtmlTag
-{
-    public Token ToHtml(string value)
-    {
-        throw new NotImplementedException();
-    }
-}
