+++ cs/Markdown/Html/Tags/BoldTag.cs
@@ -1,0 +1,9 @@
+﻿namespace Markdown.Html.Tags;
+
+public class BoldTag : IHtmlTag
+{
+    public bool TryConvert(string content, bool isClosedTag, out string resultContent)
+    {
+        throw new NotImplementedException();
+    }
+}
