+++ cs/Markdown/Html/Tags/HeaderTag.cs
@@ -1,0 +1,9 @@
+﻿namespace Markdown.Html.Tags;
+
+public class HeaderTag : IHtmlTag
+{
+    public bool TryConvert(string content, bool isClosedTag, out string resultContent)
+    {
+        throw new NotImplementedException();
+    }
+}
