+++ cs/Markdown/Html/Tags/IHtmlTag.cs
@@ -1,0 +1,6 @@
+﻿namespace Markdown.Html.Tags;
+
+public interface IHtmlTag
+{
+    bool TryConvert(string content, bool isClosedTag, out string resultContent);
+}
